Conflicts:
	ImportRenewals/Connectors/Cisco/ImportCsv.cs

+++ ImportRenewals/Models/Company.cs
@@ -9,7 +9,7 @@ namespace ImportRenewals.Models
     public class Company
     {
         [Key]
-        public Int32 CompanyId { get; set; }
+        public Int64 CompanyId { get; set; }
         public string Hash { get; set; }
         public string WestconKey { get; set; }//Para os casos do arquivo DePara, representa o Be GEO ID
         public string VendorKey { get; set; }//Para os casos do arquivo DePara, representa o Be ID
